study/Problem-1-DataSynchronization/requirements.txt
pyspark
kafka-python
mysql-connector-python
python-dotenv
